safer log

+++ LLSimpleCameraExample/ImageViewController.m
@@ -42,7 +42,7 @@
     self.imageView.image = self.image;
     [self.view addSubview:self.imageView];
     
-    NSString *info = [NSString stringWithFormat:@"Size: %@  -  Orientation: %d", NSStringFromCGSize(self.image.size), self.image.imageOrientation];
+    NSString *info = [NSString stringWithFormat:@"Size: %@  -  Orientation: %ld", NSStringFromCGSize(self.image.size), (long)self.image.imageOrientation];
     
     self.infoLabel = [[UILabel alloc] initWithFrame:CGRectMake(0, 0, 100, 20)];
     self.infoLabel.backgroundColor = [[UIColor darkGrayColor] colorWithAlphaComponent:0.7];
